CAS-180 Rename context extension class
